rename re_renderer's video decoder to video_player
